// ui/src/redux/services/types.ts
export type Movie = {
  title: string;
  posterUrl: string;
  releaseYear: number;
  description: string;
  genre: string;
  id: string;
  rating: number;
  director: string;
  reviewIds: string[];
};

export type Cinema = {
  id: string;
  name: string;
  movieIds: string[];
};

export type Review = {
  id: string;
  name: string;
  text: string;
  rating: number;
};